requirements.txt
streamlit
python-dotenv
google-generativeai
youtube-transcript-api
langchain
langchain_community
langchain-google-genai
langchain-experimental
sentence-transformers
faiss-cpu  # Using FAISS instead of ChromaDB
langchainhub
pypdf
rapidocr-onnxruntime
PyPDF2
pysqlite3-binary  # Ensures SQLite 3.35+
